Update nvim config

+++ .config/nvim/sonictemplate/cpp/base-atcoder.cpp
@@ -1,6 +1,6 @@
 #include <bits/stdc++.h>
 using namespace std;
-#define rep(i, n) for (int i = 0; i < (n); i++)
+#define rep(i, n) for (int i = 0; i < (int)(n); i++)
 using ll = long long;
 
 int main() {
